azure: Move ID parsing from `azuretasks` to `azure` utils
